refactor : 채팅룸 생성 로직 리팩토링
